sql bug

+++ utils/sql.py
@@ -53,7 +53,7 @@ FROM usage_detail_rows
 COMPUTE_CREDITS_PER_DAY = '''
 SELECT
 DATE_TRUNC('day', convert_timezone('UTC',start_time))::DATE as usage_week
-,ROUND(SUM(credits_used_compute),0) AS "Compute Credits Used"
+,ROUND(SUM(credits_used_compute),0)::FLOAT AS "Compute Credits Used"
 FROM snowflake.organization_usage.warehouse_metering_history
 WHERE start_time BETWEEN date_trunc('day', dateadd('day',-365,convert_timezone('UTC',current_timestamp()))) AND current_timestamp()
 GROUP BY 1
